fix: resolve firefox bugs (and hopefully safari)
First problem is that optional chaining wasn't supported in the browser, so removed the only instance.

Second problem was that searchParams.entries something that has a "reduce" function defined in chrome, but not in firefox (and probably safari). It's an iterable, so using "Array.from" to convert it to an array gives access to reduce again.